requirements.txt
Django==1.9.4
mock==1.3.0
python-instagram==1.3.2
pytz==2015.7
